Split unit tests into a series of smaller files
