+++ test/github.js
@@ -1,0 +1,6 @@
+/* globals describe */
+
+describe('github', function () {
+  require('./repo-github')
+  require('./repo-other')
+})
